chore(releasing): add release tooling: standard-version, commitlint, etc.
The objective is to streamline releaases and promote an enjoyable and easy open
source environment.

Commits must now follow a pre-defined format, that allows us to quickly make
releases with an up-to-date Changelog.

Tools and configuration added:
 - commitizen
 - commitlint
 - husky
 - standard-version
 - prettier
 - lint-staged

Another change made here is that we try to enforce the usage of Prettier for code formatting.

Contribution guidelines have seen a slight update.

Finally Github issue and PR templates were added.