fix(game): increase estimated game finshed time
Set extra duration for the inter-stage sync buffer to 24h, to avoid the estimated game finshed time
to ever be before than the actual finished time. This would make the game would stop progress, which
would happen with many rounds, and therefor a lot of buffer inter-stage time added to the estimated
total time. We recognize at this point that the estimated finished time field on game is probably
unnecessary and should be removed in a future update.

Fix #47