fix(exitsteps): avoid throwing if onSubmit is called without data
This error was not impacting behavior, but when onSubmit was called without data, it would try to
encode a circular reference object (a React SyntheticEvent) that was passed in the place of the data
(React automatic appending of args...) and fail to do so. But the step was still registered as done
and the bad data never recorded and it did not break the UI. So the whole issue was a noop, but at
least it doesn't throw in the console anymore.