perf: improve performance issues on games with many rounds/stages
Each round and stage for a game was loaded to the browser for access from the front end code. While
that is convenient, it is unsustainable when there are many rounds/stages in a game.

BREAKING CHANGE: Experiments depending on all rounds/stages being loaded in browser during a game
will break. It is no longer possible to have such access in the browser, but it remains available in
the server callbacks, where any logic depending on rounds/stages other than the current ones is
recommended to move.

fixes #62